Start moving things to RobotConfig
